Enable update pricingTier when update webapp
