+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/HandlerFactoryImpl.java
@@ -16,6 +16,17 @@ import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.WebAppConfiguration;
 import com.microsoft.azure.maven.webapp.configuration.DockerImageType;
 import com.microsoft.azure.maven.webapp.configuration.SchemaVersion;
+import com.microsoft.azure.maven.webapp.handlers.artifact.ArtifactHandlerImplV2;
+import com.microsoft.azure.maven.webapp.handlers.artifact.JarArtifactHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.artifact.NONEArtifactHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.artifact.WarArtifactHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.BaseRuntimeHandler;
+import com.microsoft.azure.maven.webapp.handlers.runtime.LinuxRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.NullRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PrivateDockerHubRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PrivateRegistryRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PublicDockerHubRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.WindowsRuntimeHandlerImpl;
 import com.microsoft.azure.maven.webapp.utils.WebAppUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
