+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/RuntimeHandler.java
@@ -6,6 +6,7 @@
 
 package com.microsoft.azure.maven.webapp.handlers;
 
+import com.microsoft.azure.management.appservice.AppServicePlan;
 import com.microsoft.azure.management.appservice.WebApp;
 import com.microsoft.azure.management.appservice.WebApp.DefinitionStages.WithCreate;
 import com.microsoft.azure.management.appservice.WebApp.Update;
@@ -14,4 +15,6 @@ public interface RuntimeHandler {
     WithCreate defineAppWithRuntime() throws Exception;
 
     Update updateAppRuntime(final WebApp app) throws Exception;
+
+    AppServicePlan updateAppServicePlan(final WebApp app) throws Exception;
 }
