+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/runtime/PublicDockerHubRuntimeHandlerImpl.java
@@ -4,28 +4,27 @@
  * license information.
  */
 
-package com.microsoft.azure.maven.webapp.handlers;
+package com.microsoft.azure.maven.webapp.handlers.runtime;
 
 import com.microsoft.azure.management.appservice.AppServicePlan;
 import com.microsoft.azure.management.appservice.OperatingSystem;
 import com.microsoft.azure.management.appservice.WebApp;
-import com.microsoft.azure.management.appservice.WebApp.Update;
 import com.microsoft.azure.maven.webapp.utils.WebAppUtils;
 
-public class LinuxRuntimeHandlerImpl extends BaseRuntimeHandler {
-    public static class Builder extends BaseRuntimeHandler.Builder<Builder>{
+public class PublicDockerHubRuntimeHandlerImpl extends BaseRuntimeHandler {
+    public static class Builder extends BaseRuntimeHandler.Builder<PublicDockerHubRuntimeHandlerImpl.Builder>{
         @Override
-        protected Builder self() {
+        protected PublicDockerHubRuntimeHandlerImpl.Builder self() {
             return this;
         }
 
         @Override
-        public LinuxRuntimeHandlerImpl build() {
-            return new LinuxRuntimeHandlerImpl(this);
+        public PublicDockerHubRuntimeHandlerImpl build() {
+            return new PublicDockerHubRuntimeHandlerImpl(this);
         }
     }
 
-    private LinuxRuntimeHandlerImpl(final Builder builder) {
+    private PublicDockerHubRuntimeHandlerImpl(final PublicDockerHubRuntimeHandlerImpl.Builder builder) {
         super(builder);
     }
 
@@ -34,13 +33,13 @@ public class LinuxRuntimeHandlerImpl extends BaseRuntimeHandler {
         final AppServicePlan plan = WebAppUtils.createOrGetAppServicePlan(servicePlanName, resourceGroup, azure,
             servicePlanResourceGroup, region, pricingTier, log, OperatingSystem.LINUX);
         return WebAppUtils.defineLinuxApp(resourceGroup, appName, azure, plan)
-            .withBuiltInImage(runtime);
+            .withPublicDockerHubImage(image);
     }
 
     @Override
-    public Update updateAppRuntime(WebApp app) throws Exception {
+    public WebApp.Update updateAppRuntime(final WebApp app) throws Exception {
         WebAppUtils.assureLinuxWebApp(app);
         WebAppUtils.clearTags(app);
-        return app.update().withBuiltInImage(runtime);
+        return app.update().withPublicDockerHubImage(image);
     }
 }
