+++ azure-webapp-maven-plugin/src/main/resources/com/microsoft/azure/maven/webapp/handlers/artifact/web.config.template
