+++ azure-webapp-maven-plugin/src/test/java/com/microsoft/azure/maven/webapp/handlers/HandlerFactoryImplTest.java
@@ -17,6 +17,14 @@ import com.microsoft.azure.maven.artifacthandler.ArtifactHandler;
 import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.WebAppConfiguration;
 import com.microsoft.azure.maven.webapp.configuration.OperatingSystemEnum;
+import com.microsoft.azure.maven.webapp.handlers.artifact.JarArtifactHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.artifact.WarArtifactHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.LinuxRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.NullRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PrivateDockerHubRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PrivateRegistryRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.PublicDockerHubRuntimeHandlerImpl;
+import com.microsoft.azure.maven.webapp.handlers.runtime.WindowsRuntimeHandlerImpl;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.project.MavenProject;
